pyproject.toml
[project]
name = "find-product-info"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.9,<3.11"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "openai (>=1.76.0,<2.0.0)",
    "devtools (>=0.12.2,<0.13.0)",
    "googlesearch-python (>=1.3.0,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
